requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
marshmallow==3.19.0
newsapi-python==0.2.7
psycopg2==2.9.5
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.14
